LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/canvas/element/text/2d.text.measure.emHeights-zero-descent-expected.txt

2d.text.measure.emHeights-zero-descent
Testing emHeights with zero descent metric
Actual output:

FAIL Testing emHeights with zero descent metric assert_equals: ctx.measureText('A').emHeightAscent === 40 (got 69[number], expected 40[number]) expected 40 but got 69
